Carry the voice gain chain in double, as upstream does
Three narrowings on the path from a voice to the output buffers, found
while chasing what reaches the insertion bus.  Upstream keeps every one
of these as an ordinary number and rounds only where a Float32Array is
written; the port rounded at each step.

  - Both pan tables were built with cosf/sinf from a float constant and
    a float division.  Upstream's tables are Float32Array, but each
    entry is computed with Math.cos and rounded once on the store.
  - The gain chain in render_voice: voice gain, output gain, the pan sum
    and index, the left/right gains, and the reverb, chorus and delay
    send gains.
  - The volume envelope's per-block gain ramp, and the ramp the lowpass
    filter applies with it.

No file moves much, because each is worth a bit or two per sample: STORM
105.69 -> 105.78, FOLK2 85.16 -> 85.17, the corpus unchanged at 1
very-close and 30 near-identical.  They are still what upstream does,
and they are what the insertion bus receives.

+++ spessasynth_core/include/spessasynth/synthesizer/synth.h
@@ -92,7 +92,7 @@ void SPESSASYNTH_EXPORTS ss_lowpass_filter_apply(SS_LowpassFilter *f,
                                                  const int16_t *modulated_generators,
                                                  float *buffer, int count,
                                                  float fc_excursion, float smoothing,
-                                                 float gain, float gain_inc);
+                                                 double gain, double gain_inc);
 
 /* ── Volume envelope ─────────────────────────────────────────────────────── */
